fix: suppress QuickJS and platform-conditional compiler warnings
- Add explicit (int32_t)/(double) casts in quickjs.h JS_NewInt64/JS_NewUint32
  to silence -Wconversion/-Wsign-conversion in C++ translation units
- Broaden CMake warning suppression for QuickJS C sources: add -Wno-conversion,
  -Wno-sign-conversion, -Wno-shorten-64-to-32, -Wno-float-conversion, -Wno-shadow
- Extend pragma diagnostic guards in quickjs_host.hpp for both Clang and GCC:
  add -Wconversion, -Wfloat-conversion, -Wshadow suppressions
- Add -Wno-conversion/-Wno-float-conversion/-Wno-shadow to quickjs_host.cpp flags
- Add (void) casts for unused max_depth/max_children_per_node in
  ui_automation_explorer.cpp #else fallback branch

+++ external/quickjs/quickjs.h
@@ -585,9 +585,9 @@ static js_force_inline JSValue JS_NewInt64(JSContext *ctx, int64_t val)
 {
     JSValue v;
     if (val == (int32_t)val) {
-        v = JS_NewInt32(ctx, val);
+        v = JS_NewInt32(ctx, (int32_t)val);
     } else {
-        v = __JS_NewFloat64(ctx, val);
+        v = __JS_NewFloat64(ctx, (double)val);
     }
     return v;
 }
@@ -596,7 +596,7 @@ static js_force_inline JSValue JS_NewUint32(JSContext *ctx, uint32_t val)
 {
     JSValue v;
     if (val <= 0x7fffffff) {
-        v = JS_NewInt32(ctx, val);
+        v = JS_NewInt32(ctx, (int32_t)val);
     } else {
         v = __JS_NewFloat64(ctx, val);
     }
